Make waveletdiff's derivative an FFT, and stop integer input truncating derivatives
Phi and Phi_prime are both circulant, so the DFT diagonalizes them and Phi_prime @ Phi^-1 becomes multiplication by the single transfer function fft(phi')/fft(phi), replacing a sparse LU on a (3N-2)-square matrix with an FFT pair: 20-41x faster and identical to 5e-13. Five methods allocated outputs with np.empty_like, inheriting an integer input's dtype and silently truncating derivatives, and the five taking dt_or_t now require it strictly increasing rather than accepting reversed or repeated timestamps. polydiff hands polyfit sqrt(kernel), since polyfit weights the unsquared residual, so a sample weighs the same fitting a window as blending windows; its test settings move to degree 5 and window 15, tightening five bounds by a decade.

+++ pynumdiff/linear_model.py
@@ -152,7 +152,7 @@ def lineardiff(x, dt, params=None, options=None, order=None, gamma=None, window_
     x_work = np.moveaxis(x, axis, 0) # differentiation axis to front
     shape = x_work.shape             # remember it to restore the input's dimensionality
     x_flat = x_work.reshape(shape[0], -1) # rest of the dims flattened into columns
-    x_hat = np.empty(x_flat.shape); dxdt_hat = np.empty(x_flat.shape) # not empty_like, which would inherit an integer dtype and truncate
+    x_hat = np.empty(x_flat.shape, dtype=float); dxdt_hat = np.empty(x_flat.shape, dtype=float) # float explicitly, so inherited integer input type cannot silently truncate
 
     for i in range(x_flat.shape[1]):
         v = x_flat[:, i]
